Model implementation phase two
Adapted the script and folder structure so it can be run on the Delft Blue supercomputer

Vectorized g_storage_reliability

Created a method that keeps track of the month to calculate the monthly average for each objective.

Created a method that calculates the daily hydropower generation from the six release decisions being taken on a daily basis.

Since the recreational objective depends on a Boolean statement that is not true for all months, we decided to repeat the average yearly recreational reliability 12*times.

Added standard deviation method that calculates standard deviation out of an array. Specifically the purpose is to calculate the standard deviation between the objectives stored on a monthly basis.

Changed the hydropower objective by calculating the total energy produced multiplied by the fraction of energy generated from the Conowingo dam.

Added the rawlsian difference principle, and the sufficientarian principle in the objective formulation.

Issue #2  is ongoing, namely how to store the results. Currently I get a dataframe as output with only one row. Ideally, the row should be extended to the length of the amount of solutions found.